For purposes of naming, switch data sizes to uint32_t to support full codes. I am unsure why mfgCode was used to have a cluster code, but it seems to behave like that

+++ src/app/util/util.h
@@ -74,8 +74,8 @@
 // Cluster name structure
 typedef struct
 {
-    uint16_t id;
-    uint16_t mfgCode;
+    uint32_t id;
+    uint32_t mfgCode;
     const char * name;
 } EmberAfClusterName;
